--- core/client/components/router/spa-router.js
@@ -1,57 +1,0 @@
-import { useActionState, useEffect, startTransition } from 'react';
-import { RouterContext } from './index.js';
-import { routerReducer } from '../../lib/router-reducer.js';
-import { getFullPath } from '../../lib/utils.js';
-import { FlashProvider } from '../flash-provider.js';
-import { SlotContext } from '../slot-provider.js';
-
-export function SPARouter({ initialState, initialFlashMessages, rootLayout }) {
-  const [routerState, dispatch, isTransitioning] = useActionState(
-    routerReducer,
-    initialState
-  );
-
-  useEffect(() => {
-    const controller = new AbortController();
-
-    const updateTree = (tree) => {
-      startTransition(() => {
-        dispatch({ type: 'UPDATE', payload: { tree } });
-      });
-    };
-
-    window.__cosmos_rsc = { ...window.__cosmos_rsc, updateTree };
-
-    const handlePopState = () => {
-      startTransition(() => {
-        dispatch({
-          type: 'NAVIGATE',
-          payload: { url: getFullPath(window.location.href) },
-        });
-      });
-    };
-
-    window.addEventListener('popstate', handlePopState, {
-      signal: controller.signal,
-    });
-
-    return () => {
-      controller.abort();
-      delete window.__cosmos_rsc.updateTree;
-    };
-  }, [dispatch]);
-
-  const push = (url) => {
-    startTransition(() => {
-      dispatch({ type: 'PUSH', payload: { url: getFullPath(url) } });
-    });
-  };
-
-  return (
-    <RouterContext value={{ push, isTransitioning }}>
-      <FlashProvider initialState={initialFlashMessages}>
-        <SlotContext value={routerState.tree}>{rootLayout}</SlotContext>
-      </FlashProvider>
-    </RouterContext>
-  );
-}
