+++ core/client/exports.js
@@ -3,5 +3,5 @@
 import 'client-only';
 
 export { Link } from './components/link.js';
-export { useRouter } from './components/router/index.js';
-export { useFlash } from './components/flash-provider.js';
+export { useRouter } from './components/router-context.js';
+export { useFlash } from './components/flash-context.js';
