+++ core/client/index.js
@@ -2,7 +2,7 @@ import { hydrateRoot } from 'react-dom/client';
 import { createFromReadableStream } from 'react-server-dom-webpack/client';
 import { ErrorBoundary } from './components/error-boundary';
 import { rscStream } from '../rsc-html-stream/client';
-import { SPARouter } from './components/router/spa-router';
+import { BrowserApp } from './components/app/browser-app';
 import { callServer } from './lib/call-server';
 import { getFullPath } from './lib/utils';
 import { StrictMode } from 'react';
@@ -15,24 +15,21 @@ async function hydrateDocument() {
 
   const initialState = {
     tree,
+    flashMessages,
     cache: new Map([[getFullPath(window.location.href), tree]]),
   };
 
-  hydrateRoot(
-    document,
+  const app = (
     <StrictMode>
       <ErrorBoundary>
-        <SPARouter
-          rootLayout={rootLayout}
-          initialState={initialState}
-          initialFlashMessages={flashMessages}
-        />
+        <BrowserApp rootLayout={rootLayout} initialState={initialState} />
       </ErrorBoundary>
-    </StrictMode>,
-    {
-      formState,
-    }
+    </StrictMode>
   );
+
+  hydrateRoot(document, app, {
+    formState,
+  });
 }
 
 hydrateDocument();
