+++ core/server/loaders/jsx.js
